# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reparto API",
    "version": "0.2.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/clientes/buscar": {
      "get": {
        "summary": "Autocompletado de clientes por razón social o dirección (tolera tildes y errores de dictado)",
        "tags": [
          "clientes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "in": "query",
            "name": "q",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "in": "query",
            "name": "comuna",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            },
            "in": "query",
            "name": "limite",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultados": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "localId": {
                            "type": "string"
                          },
                          "clienteId": {
                            "type": "string"
                          },
                          "razonSocial": {
                            "type": "string"
                          },
                          "direccion": {
                            "type": "string"
                          },
                          "comuna": {
                            "type": "string"
                          },
                          "lat": {
                            "type": "number"
                          },
                          "lng": {
                            "type": "number"
                          },
                          "pinEstado": {
                            "type": "string",
                            "enum": [
                              "pendiente",
                              "sugerido",
                              "validado"
                            ]
                          },
                          "fotoPath": {
                            "type": "string"
                          },
                          "streetviewRumbo": {
                            "type": "number"
                          },
                          "nota": {
                            "type": "string"
                          },
                          "score": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "localId",
                          "clienteId",
                          "razonSocial",
                          "direccion",
                          "comuna",
                          "pinEstado",
                          "score"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "resultados"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/clientes": {
      "post": {
        "summary": "Cliente nuevo con su primer local",
        "tags": [
          "clientes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rut": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "razonSocial": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "giro": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "direccion": {
                    "type": "string",
                    "maxLength": 600
                  },
                  "comuna": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "lat": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "string",
                        "maxLength": 40
                      }
                    ]
                  },
                  "lng": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "string",
                        "maxLength": 40
                      }
                    ]
                  },
                  "nota": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clienteId": {
                      "type": "string"
                    },
                    "localId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "clienteId",
                    "localId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}